* Move `fluxTransform()` body to `Transformers` for cleaner context distribution
