+++ spring-integration-core/src/main/java/org/springframework/integration/dsl/IntegrationFlowDefinition.java
@@ -22,7 +22,6 @@ import java.util.Map;
 import java.util.Optional;
 import java.util.Set;
 import java.util.concurrent.Executor;
-import java.util.concurrent.atomic.AtomicReference;
 import java.util.function.Consumer;
 import java.util.function.Function;
 
@@ -80,7 +79,6 @@ import org.springframework.integration.splitter.ExpressionEvaluatingSplitter;
 import org.springframework.integration.splitter.MethodInvokingSplitter;
 import org.springframework.integration.store.MessageStore;
 import org.springframework.integration.support.MapBuilder;
-import org.springframework.integration.support.MessageBuilder;
 import org.springframework.integration.transformer.ClaimCheckInTransformer;
 import org.springframework.integration.transformer.ClaimCheckOutTransformer;
 import org.springframework.integration.transformer.ContentEnricher;
@@ -99,7 +97,6 @@ import org.springframework.util.CollectionUtils;
 import org.springframework.util.StringUtils;
 
 import reactor.core.publisher.Flux;
-import reactor.core.publisher.Mono;
 import reactor.util.function.Tuple2;
 
 /**
@@ -2885,28 +2882,10 @@ public abstract class IntegrationFlowDefinition<B extends IntegrationFlowDefinit
 
 		Publisher<Message<I>> upstream = (Publisher<Message<I>>) this.currentMessageChannel;
 
-		Flux<Message<?>> result =
-				Flux.from(upstream)
-						.flatMap(message ->
-								Mono.subscriberContext()
-										.map(ctx -> {
-											ctx.get(RequestMessageHolder.class).set(message);
-											return message;
-										}))
-						.transform(fluxFunction)
-						.flatMap(data ->
-								data instanceof Message<?>
-										? Mono.just((Message<?>) data)
-										: Mono.subscriberContext()
-												.map(ctx -> ctx.get(RequestMessageHolder.class).get())
-												.map(requestMessage ->
-														MessageBuilder.withPayload(data)
-																.copyHeaders(requestMessage.getHeaders())
-																.build()))
-						.subscriberContext(ctx -> ctx.put(RequestMessageHolder.class, new RequestMessageHolder()));
+		Flux<Message<O>> result = Transformers.transformWithFunction(upstream, fluxFunction);
 
 		FluxMessageChannel downstream = new FluxMessageChannel();
-		downstream.subscribeTo(result);
+		downstream.subscribeTo((Flux<Message<?>>) (Flux<?>) result);
 
 		this.currentMessageChannel = downstream;
 
@@ -3131,10 +3110,6 @@ public abstract class IntegrationFlowDefinition<B extends IntegrationFlowDefinit
 		public void postProcessBeforeDestruction(Object bean, String beanName) throws BeansException {
 			IntegrationFlowDefinition.REFERENCED_REPLY_PRODUCERS.remove(bean);
 		}
-
-	}
-
-	private static class RequestMessageHolder extends AtomicReference<Message<?>> {
 
 	}
