+++ spring-integration-core/src/main/java/org/springframework/integration/dsl/Transformers.java
@@ -16,7 +16,10 @@
 
 package org.springframework.integration.dsl;
 
+import java.util.concurrent.atomic.AtomicReference;
 import java.util.function.Function;
+
+import org.reactivestreams.Publisher;
 
 import org.springframework.core.convert.converter.Converter;
 import org.springframework.core.serializer.Deserializer;
@@ -27,6 +30,7 @@ import org.springframework.integration.codec.Codec;
 import org.springframework.integration.expression.FunctionExpression;
 import org.springframework.integration.json.JsonToObjectTransformer;
 import org.springframework.integration.json.ObjectToJsonTransformer;
+import org.springframework.integration.support.MessageBuilder;
 import org.springframework.integration.support.json.JsonObjectMapper;
 import org.springframework.integration.transformer.DecodingTransformer;
 import org.springframework.integration.transformer.EncodingPayloadTransformer;
@@ -40,6 +44,9 @@ import org.springframework.integration.transformer.StreamTransformer;
 import org.springframework.integration.transformer.SyslogToMapTransformer;
 import org.springframework.messaging.Message;
 import org.springframework.util.Assert;
+
+import reactor.core.publisher.Flux;
+import reactor.core.publisher.Mono;
 
 /**
  * An utility class to provide methods for out-of-the-box
@@ -260,6 +267,35 @@ public abstract class Transformers {
 	 */
 	public static StreamTransformer fromStream(String charset) {
 		return new StreamTransformer(charset);
+	}
+
+
+	@SuppressWarnings("unchecked")
+	static <I, O> Flux<Message<O>> transformWithFunction(Publisher<Message<I>> publisher,
+			Function<? super Flux<Message<I>>, ? extends Publisher<O>> fluxFunction) {
+
+		return Flux.from(publisher)
+				.flatMap(message ->
+						Mono.subscriberContext()
+								.map(ctx -> {
+									ctx.get(RequestMessageHolder.class).set(message);
+									return message;
+								}))
+				.transform(fluxFunction)
+				.flatMap(data ->
+						data instanceof Message<?>
+								? Mono.just((Message<O>) data)
+								: Mono.subscriberContext()
+										.map(ctx -> ctx.get(RequestMessageHolder.class).get())
+										.map(requestMessage ->
+												MessageBuilder.withPayload(data)
+														.copyHeaders(requestMessage.getHeaders())
+														.build()))
+				.subscriberContext(ctx -> ctx.put(RequestMessageHolder.class, new RequestMessageHolder()));
+	}
+
+	private static class RequestMessageHolder extends AtomicReference<Message<?>> {
+
 	}
 
 }
